+++ src/Http/Controllers/Admin/ModuleController.php
@@ -26,6 +26,8 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 abstract class ModuleController extends Controller
 {
+    use Concerns\FormSubmitOptions;
+
     /**
      * @var Application
      */
@@ -249,6 +251,13 @@ abstract class ModuleController extends Controller
      * @var array
      */
     protected $fieldsPermissions = [];
+
+    /**
+     * Determines if draft revisions can be added on top of published content.
+     *
+     * @var bool
+     */
+    protected $enableDraftRevisions = false;
 
     public function __construct(Application $app, Request $request)
     {
@@ -476,7 +485,17 @@ abstract class ModuleController extends Controller
             return View::exists($view);
         });
 
-        return View::make($view, $this->form($id));
+        $item = $this->repository->getById($id);
+
+        if ($this->moduleHas('revisions')) {
+            $latestRevision = $item->revisions->first();
+
+            if ($latestRevision && $latestRevision->isDraft()) {
+                Session::flash('status', twillTrans('twill::lang.publisher.draft-revisions-available'));
+            }
+        }
+
+        return View::make($view, $this->form($id, $item));
     }
 
     /**
@@ -536,13 +555,7 @@ abstract class ModuleController extends Controller
                 [Str::singular($this->moduleName) => $id]
             ));
         } else {
-            $formRequest = $this->validateFormRequest();
-
-            $this->repository->update($id, $formRequest->all());
-
-            activity()->performedOn($item)->log('updated');
-
-            $this->fireEvent();
+            $this->performUpdate($item);
 
             if (isset($input['cmsSaveType'])) {
                 if (Str::endsWith($input['cmsSaveType'], '-close')) {
@@ -578,11 +591,29 @@ abstract class ModuleController extends Controller
                 return Response::json([
                     'message' => twillTrans('twill::lang.publisher.save-success'),
                     'variant' => FlashLevel::SUCCESS,
-                    'revisions' => $item->revisionsArray(),
+                    'revisions' => $item->refresh()->revisionsArray(),
                 ]);
             }
 
             return $this->respondWithSuccess(twillTrans('twill::lang.publisher.save-success'));
+        }
+    }
+
+    protected function performUpdate($item)
+    {
+        $formRequest = $this->validateFormRequest();
+        $data = $formRequest->all();
+
+        if (Str::startsWith($data['cmsSaveType'] ?? '', 'draft-revision')) {
+            $data['published'] = false;
+
+            $this->repository->createRevisionIfNeeded($item, $data);
+        } else {
+            $this->repository->update($item->id, $data);
+
+            activity()->performedOn($item)->log('updated');
+
+            $this->fireEvent();
         }
     }
 
@@ -637,9 +668,13 @@ abstract class ModuleController extends Controller
         });
 
         $revision = $item->revisions()->where('id', $this->request->get('revisionId'))->first();
-        $date = $revision->created_at->toDayDateTimeString();
 
-        Session::flash('restoreMessage', twillTrans('twill::lang.publisher.restore-message', ['user' => $revision->byUser, 'date' => $date]));
+        if ($revision->isDraft()) {
+            Session::flash('restoreMessage', twillTrans('twill::lang.publisher.editing-draft-revision'));
+        } else {
+            $date = $revision->created_at->toDayDateTimeString();
+            Session::flash('restoreMessage', twillTrans('twill::lang.publisher.restore-message', ['user' => $revision->byUser, 'date' => $date]));
+        }
 
         return View::make($view, $this->form($id, $item));
     }
@@ -1558,6 +1593,7 @@ abstract class ModuleController extends Controller
             'blockPreviewUrl' => Route::has('admin.blocks.preview') ? URL::route('admin.blocks.preview') : '#',
             'availableRepeaters' => $this->getRepeaterList()->toJson(),
             'revisions' => $this->moduleHas('revisions') ? $item->revisionsArray() : null,
+            'submitOptions' => $this->getSubmitOptions($item),
         ] + (Route::has($previewRouteName) && $itemId ? [
             'previewUrl' => moduleRoute($this->moduleName, $this->routePrefix, 'preview', [$itemId]),
         ] : [])